packages/ai/requirements.txt
structlog==24.1.0
eth-typing==4.2.1
web3==6.19.0
pytest==8.2.2
pytest-asyncio==0.23.7
pydantic-settings==2.3.4
httpx==0.27.0
python-dotenv==1.0.1
google-generativeai==0.5.4
cryptography==42.0.7
PyJWT==2.8.0
pydantic-ai==0.0.18
fastapi==0.111.0
uvicorn==0.30.1
